feat(app): add tab-chrome command center, enable minimap
- Add command_center_rect shared by painting and hit testing so Quick
  Open is discoverable within the tab chrome without consuming another
  permanent row of vertical editor space
- Enable the minimap by default in Settings and update the README
  example to match
- Retitle the window to "LightLine IDE"
- Refactor the format_on_save round-trip test to use struct update
  syntax
- Replace map_or(false, ...) with is_some_and and simplify the theme
  composite helper to satisfy clippy lints

+++ src/windows_app/color_theme_adapter.rs
@@ -23,9 +23,9 @@ use lightline::color_theme::{Rgba, ZedColorTheme};
 // (much too vivid) foreground color at full strength.
 fn composite(color: Rgba, backdrop: u32) -> u32 {
     let a = u32::from(color.a);
-    let br = (backdrop & 0xff) as u32;
-    let bg = ((backdrop >> 8) & 0xff) as u32;
-    let bb = ((backdrop >> 16) & 0xff) as u32;
+    let br = backdrop & 0xff;
+    let bg = (backdrop >> 8) & 0xff;
+    let bb = (backdrop >> 16) & 0xff;
     let blend = |fg: u8, bd: u32| -> u32 { (u32::from(fg) * a + bd * (255 - a)) / 255 };
     blend(color.r, br) | (blend(color.g, bg) << 8) | (blend(color.b, bb) << 16)
 }
